+++ src/types/IReactComponent.ts
@@ -1,0 +1,4 @@
+export type IReactComponent<P = any> =
+    | React.ClassicComponentClass<P>
+    | React.ComponentClass<P>
+    | React.StatelessComponent<P>
