+++ src/types/IStoresToProps.ts
@@ -1,0 +1,7 @@
+import { IValueMap } from "./IValueMap"
+export type IStoresToProps<
+    S extends IValueMap = {},
+    P extends IValueMap = {},
+    I extends IValueMap = {},
+    C extends IValueMap = {}
+> = (stores: S, nextProps: P, context?: C) => I
