+++ src/types/IValueMap.ts
@@ -1,0 +1,1 @@
+export type IValueMap = Record<string, any>
