+++ src/types/IWrappedComponent.ts
@@ -1,0 +1,4 @@
+import { IReactComponent } from "./IReactComponent"
+export type IWrappedComponent<P> = {
+    wrappedComponent: IReactComponent<P>
+}
